Fix issue preventing opt_main --pipeline_proto from working.
A while ago a feature was added to allow the runtime replacement of pass pipeline protos (which define the available compound passes and the default pipeline) at runtime. This was not tested particularly well and had several issues. Most notable the flag which allows one to use this feature was buggy and did not work and the redefined compound passes were not used in all circumstances, sometimes falling back to the built-in passes.

These issues have been fixed by the last 2 changes and better quality tests have been added.

PiperOrigin-RevId: 844913821