Ensure --skip_passes does not care about configured min-opt-level etc.
Several passes are (conditionally) disabled based on the opt level. This led to --skip_passes to behave in surprising ways since to skip a pass you need to include any of the configurations too and which ones are needed depends on the exact specifics of where in the pipeline the pass is declared.

This makes everything simpler by forcing a pass to be skipped if its base short_name matches the argument.

This does prevent you from specifically disabling (eg) just 'dfe(opt_level>=2)' or whatever but this feature is not often used anyway and the behavior of 'dfe' just disabling all dead-function runs regardless of position is more likely to be the actually desired result.

PiperOrigin-RevId: 844945459