[xls][mlir] Apply IoConstraints to make XLS scheduler respect # pipeline stage.
This CL adds an  IoConstraints attribute to EprocOp that is used as io_constraint
in the XLS codegen options. IO constraints are added to eprocs in the proc-
elaboration pass if `use_io_constraints_on_eprocs` is set.

If constraints are specified we increase the  `pipline_stages` for the XLS scheduler
by 2 to account for the input- and output-flopping. We observed that XLS scheduler
failed when trying to use the same number of pipeline stages as the IoConstraint.

PiperOrigin-RevId: 845126566