Refactor `OptimizationContext::SharedNodeData` and `SharedQueryEngine` to use variadic templates.
This lets you construct and retrieve shared data between passes with arbitrary constructor arguments. Care should be taken not to leak too much data as these are only cleaned up when optimization is finished.

Also changed the SharedNodeData to return status. This should probably be done with the query engine too at some point but that is a much larger refactor so it can wait.

PiperOrigin-RevId: 845931292