Add support for non-synthesizable functions in XLS IR.
Introduces a `#[non_synth]` attribute for functions, which marks them as not being synthesized. Adds decorator estimators (`FilterNonSynth`) for area and delay models to ignore nodes within non-synth functions during optimization. Updates the IR parser, verifier, and non-synth separation pass to handle this new attribute.

This is needed to ensure that all passes are able to cope with the existence of non-synth function invokes during the optimization process.

PiperOrigin-RevId: 845962215