Fix new FSM bug in RemovePassThroughs()
When feedback passthroughs are removed, this can change the order of ContinuationValues seen by an input phi, that is which value the phi sees in each activation, corrupting the logic of the block.

This change disallows removing feedback passthroughs in RemovePassThroughs().

This creates a QoR problem that sometimes direct-ins get stored in StateElements. This happens because feedbacks are always stored in StateElements, even if they are sourced from direct-ins. This is because, for example "value+1" is still marked direct-in as long as value is marked direct-in.

To address this, the pass RemovePassthroughFeedbacks() is added that removes feedback passthroughs only if they all feed in the same value to the phi as the feed-forward input.

Note that more comprehensive and/or elegant solutions could include:
- Doing virtual unrolling before continuation optimization
- Removing feedback passthroughs if they pass a more sophisticated analysis that they don't cross other feedbacks on the same phi.

PiperOrigin-RevId: 846479545