[scheduling] Clean up after adding slack variables for SDC scheduling errors
A new test case is added to `pipeline_scheduling_pass_test` to verify behavior when rerunning scheduling after a failure due to a tight clock period.

Along with a small patch to our error reporting, this should ensure we always give the best explanation we can, given what the user requested.

PiperOrigin-RevId: 848165775