[codegen 1.5] Gate receive data based on predicate and valid signals.
When `gate_recvs` is enabled, the data output of a receive operation is gated. The gate condition is the logical AND of the receive's predicate (if present) and the channel's valid signal (if the receive is non-blocking). If the gate condition is false, the receive data is zero'd out for a forced match with simulated behavior.

PiperOrigin-RevId: 849163226