Add a test case for a "join" Proc in block conversion.
This test case demonstrates the current behavior of converting a Proc that receives from two channels and sends to one. It highlights an issue where one of the receive-ready signals can be asserted even when the other input is not yet valid, potentially leading to data loss.

Documents google/xls#3605

PiperOrigin-RevId: 849273814